[Bronze II] Title: Hashing, Time: 124 ms, Memory: 14188 KB -BaekjoonHub

+++ 백준/Bronze/15829. Hashing/Hashing.java
@@ -2,7 +2,7 @@ import java.io.*;
 import java.util.*;
  
 public class Main {
-	static int N;
+	static long N;
 	static long result;
 	static int[] num;
 	static char[] ch;
